requirements.txt
asgiref==3.2.7
Django==3.0.7
dnspython==1.16.0
enum-compat==0.0.3
eventlet==0.25.2
greenlet==0.4.16
monotonic==1.5
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
